Move status bar item to the right
The character info should now be next to the line/column display (number value by try & error & retry same & other error...). The left side of the status bar is already pretty crowded with project actions and the right side seems more for the editor itself, so I find it more appropriate to put it there.